+++ packages/api-client/src/components/CommandPalette/CommandPaletteImport.vue
@@ -128,7 +128,7 @@ async function importCollection() {
     <template v-if="!documentDetails || isUrl(inputContent)">
       <CommandActionInput
         v-model="inputContent"
-        placeholder="Your OpenAPI/Swagger document or URL"
+        placeholder="OpenAPI/Swagger URL or document"
         @onDelete="emits('back', $event)" />
     </template>
     <template v-else>